+++ packages/xchain-bitcoin/src/utils.ts
@@ -13,6 +13,10 @@ import { FeeRate, FeeRates, FeesWithRates } from './types/client-types'
 import { BroadcastTxParams, UTXO, UTXOs } from './types/common'
 import { MIN_TX_FEE } from './const'
 
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-ignore
+import accumulative from 'coinselect/accumulative'
+
 const TX_EMPTY_SIZE = 4 + 1 + 1 + 4 //10
 const TX_INPUT_BASE = 32 + 4 + 1 + 4 // 41
 const TX_INPUT_PUBKEYHASH = 107
@@ -171,7 +175,6 @@ export const scanUTXOs = async ({
       } as UTXO),
   )
 }
-
 /**
  * Build transcation.
  *